feat: add categories to CatalogView
