+++ src/app/view/main/catalog/categories/catalog-view-types.ts
@@ -1,0 +1,14 @@
+export const CATALOG_CLASSES = {
+  CATALOG: 'catalog',
+  TITLE: 'catalog__title',
+  CONTENT: 'catalog__content',
+  ASIDE: 'catalog__content-aside',
+  PRODUCTS: 'catalog__content-products',
+};
+
+export const CATALOG_TEXT = {
+  TITLE: 'Catalog',
+  CONTENT: '',
+  ASIDE: '',
+  PRODUCTS: '',
+};
