+++ src/app/view/main/catalog/categories/filters/filters-view-types.ts
