+++ src/assets/sass/main.scss
@@ -19,6 +19,7 @@
 @import "components/cover";
 @import "components/dropdown";
 @import "components/inputs";
+@import "components/bread-crumbs";
 
 @import "pages/home";
 @import "pages/_registration";
