+++ src/assets/sass/pages/_category.scss
@@ -1,0 +1,113 @@
+// .catalog {
+//   display: flex;
+//   flex: 1;
+//   align-items: center;
+//   justify-content: center;
+//   flex-direction: column;
+//   &__title {
+//     width: 100%;
+//     text-align: center;
+//     font-size: 24px;
+//     margin: 15px 0 5px;
+//   }
+//   &__content {
+//     display: flex;
+//     border-top: 1px solid $black-color;
+//     width: 100%;
+//     flex: 1;
+//     align-items: center;
+//     justify-content: space-around;
+//     &-aside {
+//       width: 20%;
+//       border-right: 1px solid $black-color;
+//       height: 100%;
+//     }
+//     &-products {
+//       display: flex;
+//       flex: 1;
+//       flex-direction: column;
+//       flex-wrap: wrap;
+//       gap: 5px;
+//       width: 80%;
+//       .product-card {
+//         width: 200px;
+//         height: 400px;
+//         border: 1px solid $catalog-card-border-color;
+//         word-wrap: break-word;
+//         transition: ease-in-out 0.5s;
+//         padding: 7px;
+//         display: flex;
+//         gap: 5px;
+//         flex-direction: column;
+//         text-align: center;
+//         &:hover {
+//           transition: ease-in-out 0.5s;
+//           box-shadow: $catalog-card-shadow 0px 3px 8px;
+//           transform: scale(1.01);
+//           cursor: pointer;
+//         }
+//         &__name {
+//           font-size: 1.2rem;
+//         }
+//         &__image {
+//           width: auto;
+//           height: 50%;
+//           display: block;
+//           max-width: 100%;
+//           object-fit: contain;
+//         }
+//         &__price {
+//           font-size: 1.15rem;
+//           text-decoration: line-through;
+//           color: $catalog-card-old-price;
+//         }
+//         &__discount {
+//           font-size: 1.15rem;
+//         }
+//         &__description {
+//           overflow: scroll;
+//         }
+//       }
+//       .cards-wrapper {
+//         display: flex;
+//         flex-wrap: wrap;
+//         align-items: center;
+//         justify-content: center;
+//         gap: 5px;
+//         margin: 10px 0;
+//       }
+//       .paginator {
+//         gap: 5px;
+//         flex-wrap: wrap;
+//         display: flex;
+//         margin: 10px 0;
+//         align-items: center;
+//         justify-content: center;
+//         &__btn {
+//           user-select: none;
+//           position: relative;
+//           display: block;
+//           background-color: $catalog-pagination-button-background-color;
+//           border-radius: 8px;
+//           color: $catalog-pagination-button-text-color;
+//           cursor: pointer;
+//           width: 40px;
+//           height: 40px;
+//           transition: ease-in-out 0.3s;
+//           &:hover {
+//             transition: ease-in-out 0.3s;
+//             background-color: darken($catalog-pagination-button-background-color, 10%);
+//           }
+//           &.active {
+//             background-color: darken($catalog-pagination-button-background-color, 25%);
+//             cursor: auto;
+//           }
+//           &:disabled {
+//             background-color: $catalog-pagination-button-disabled-background-color;
+//             cursor: auto;
+//           }
+//         }
+//       }
+//     }
+//   }
+// }
